Resolve (#410) Logs are missing query strings (#1165)
* :white_check_mark: migrate utilities based tests from `tests.py` to `tests_utilities.py`

* :fire: remove `wsgi-request-logger` dependency (appears to no longer be maintained) and bring formatter functionality into zappa
:sparkles: add `ApacheNCSAFormatter` to replace wsgi-request-logger functionality.

* :wrench: black is no longer in 'beta', changing allow_prereleases to false.
