+++ Pipfile
@@ -40,8 +40,6 @@ tqdm = "*"
 troposphere = ">=3.0"
 Werkzeug = "*"
 wheel = "*"
-wsgi-request-logger = "*"
 
 [pipenv]
-# Required for 'black' since all of its release tags contain 'b'
-allow_prereleases = true
+allow_prereleases = false
